fixCountdown addPattern fixTimer change to ready, set, go change opacity add pattern behind game on browser added company logo fixed timer ending to register player moves as well
