+++ public/assets/js/components/Timer.js
@@ -6,8 +6,9 @@ import roundCountdown from "../helpers/roundCountdown.js";
 import clearBoardForNewRound from "../helpers/clearBoardForNewRound.js";
 import restingMode from "../helpers/restingMode.js";
 import Players from "../helpers/Players.js";
+import donePressed from "../helpers/donePressed.js";
 
-let myTimer = document.querySelector('div.timer-counter');
+// let myTimer = document.querySelector('div.timer-counter');
 let myTimerCounter = document.querySelector('span.time-nums');
 
 class Timer {
@@ -45,79 +46,81 @@ class Timer {
             this.counter--;
             // this.element.innerHTML = this.counter;
             if (this.counter === 0) {
-                const opponentMove = Players.all.player2.generateRandomMoves();
-                console.log(opponentMove)
+                donePressed()
+                
+                // const opponentMove = Players.all.player2.generateRandomMoves();
+                // console.log(opponentMove)
 
-                let opponentTallyColumn = document.querySelectorAll(`table.tally.opponent-tally td:nth-child(${Rounds.all['game1'].counter})`);
-                console.log('opponent', opponentTallyColumn);
-                opponentTallyColumn.forEach((td, index) => {
-                    let moveComponent = Object.values(opponentMove)[index];
-                    if (moveComponent === 'attack') {
-                        td.classList.add('cell-attacked');
-                    } else if (moveComponent === 'block') {
-                        td.classList.add('cell-blocked');
-                    }
-                });
+                // let opponentTallyColumn = document.querySelectorAll(`table.tally.opponent-tally td:nth-child(${Rounds.all['game1'].counter})`);
+                // console.log('opponent', opponentTallyColumn);
+                // opponentTallyColumn.forEach((td, index) => {
+                //     let moveComponent = Object.values(opponentMove)[index];
+                //     if (moveComponent === 'attack') {
+                //         td.classList.add('cell-attacked');
+                //     } else if (moveComponent === 'block') {
+                //         td.classList.add('cell-blocked');
+                //     }
+                // });
 
-                let myTallyColumn = document.querySelectorAll(`table.tally.my-tally td:nth-child(${Rounds.all['game1'].counter})`);
+                // let myTallyColumn = document.querySelectorAll(`table.tally.my-tally td:nth-child(${Rounds.all['game1'].counter})`);
 
-                if (Object.values(opponentMove).includes("attack")) {
-                    myTallyColumn.forEach(td => {
-                        td.classList.add('round-defeat');
-                    });
-                    opponentTallyColumn.forEach(td => {
-                        td.classList.add('round-won');
-                    });
-                    Life.all.myLife.decreaseCounter();
-                } else {
-                    myTallyColumn.forEach(td => {
-                        td.classList.add('round-draw');
-                    });
-                    opponentTallyColumn.forEach(td => {
-                        td.classList.add('round-draw');
-                    });
-                }
+                // if (Object.values(opponentMove).includes("attack")) {
+                //     myTallyColumn.forEach(td => {
+                //         td.classList.add('round-defeat');
+                //     });
+                //     opponentTallyColumn.forEach(td => {
+                //         td.classList.add('round-won');
+                //     });
+                //     Life.all.myLife.decreaseCounter();
+                // } else {
+                //     myTallyColumn.forEach(td => {
+                //         td.classList.add('round-draw');
+                //     });
+                //     opponentTallyColumn.forEach(td => {
+                //         td.classList.add('round-draw');
+                //     });
+                // }
 
-                if (Rounds.all['game1'].counter < Rounds.all['game1'].counterRange[1] && Life.all.myLife.counter > 0 && Life.all.opponentLife.counter > 0) {
-                    restingMode();
-                    this.resetCounter();
+                // if (Rounds.all['game1'].counter < Rounds.all['game1'].counterRange[1] && Life.all.myLife.counter > 0 && Life.all.opponentLife.counter > 0) {
+                //     restingMode();
+                //     this.resetCounter();
 
-                    setTimeout(() => {
-                        document.querySelector("div.countdown-overlay").classList.remove("d-none");
-                        roundCountdown();
-                    }, 5000)
+                //     setTimeout(() => {
+                //         document.querySelector("div.countdown-overlay").classList.remove("d-none");
+                //         roundCountdown();
+                //     }, 5000)
 
-                    setTimeout(() => {
-                        document.querySelector("div.countdown-overlay").classList.add("d-none");
-                        Rounds.all['game1'].increaseCounter();
-                        clearBoardForNewRound(Rounds.all['game1'].counter);
-                    }, 9000);
-                } else {
-                    restingMode();
-                    this.resetCounter();
-                    let resultOverlay = document.querySelector(".result-banner");
-                    let gameResult = calculateGameResults();
-                    console.log('game result', gameResult);
-                    if (gameResult === 'win') {
-                        resultOverlay.classList.add('victory');
-                    } else if (gameResult === 'loss') {
-                        resultOverlay.classList.add('defeat');
-                    } else {
-                        resultOverlay.classList.add('draw');
-                    }
-                    // resultOverlay.classList.add('victory');
-                    // resultOverlay.classList.add('draw');
-                    // resultOverlay.classList.add('defeat');
-                    // console.log(resultOverlay);
-                    let replayBtn = document.querySelector(".play-again");
-                    replayBtn.classList.add('replay-in-animation');
-                    replayBtn.classList.remove('replay-out-animation');
-                    replayBtn.classList.remove('d-none');
+                //     setTimeout(() => {
+                //         document.querySelector("div.countdown-overlay").classList.add("d-none");
+                //         Rounds.all['game1'].increaseCounter();
+                //         clearBoardForNewRound(Rounds.all['game1'].counter);
+                //     }, 9000);
+                // } else {
+                //     restingMode();
+                //     this.resetCounter();
+                //     let resultOverlay = document.querySelector(".result-banner");
+                //     let gameResult = calculateGameResults();
+                //     console.log('game result', gameResult);
+                //     if (gameResult === 'win') {
+                //         resultOverlay.classList.add('victory');
+                //     } else if (gameResult === 'loss') {
+                //         resultOverlay.classList.add('defeat');
+                //     } else {
+                //         resultOverlay.classList.add('draw');
+                //     }
+                //     // resultOverlay.classList.add('victory');
+                //     // resultOverlay.classList.add('draw');
+                //     // resultOverlay.classList.add('defeat');
+                //     // console.log(resultOverlay);
+                //     let replayBtn = document.querySelector(".play-again");
+                //     replayBtn.classList.add('replay-in-animation');
+                //     replayBtn.classList.remove('replay-out-animation');
+                //     replayBtn.classList.remove('d-none');
 
-                    document.querySelector('div.done').classList.add('d-none');
-                    document.querySelector('div.moves-placeholder').classList.add('d-none');
-                    Rounds.all['game1'].resetCounter();
-                }
+                //     document.querySelector('div.done').classList.add('d-none');
+                //     document.querySelector('div.moves-placeholder').classList.add('d-none');
+                //     Rounds.all['game1'].resetCounter();
+                // }
             }
         }, 1000);
     }
