+++ views/play.ejs
@@ -21,42 +21,41 @@
     <div class="container-fluid big-container">
         <div class="row">
             <div class="col-md-12">
-            
-            
-             <!-- Portrait Mode Overlay -->
-    <div class="flip-overlay">
-        <!-- Placeholder for incoming D3 asset -->
-        <span class="portrait">FLIP TO PLAY</span>
-    </div>
 
-    <!-- Round Countdown Overlay -->
-    <div class="countdown-overlay d-none">
-        <!-- Placeholder for incoming D3 asset -->
-        <span class="countdown">3</span>
-    </div>
+                <!-- Portrait Mode Overlay -->
+                <div class="flip-overlay">
+                    <!-- Placeholder for incoming D3 asset -->
+                    <span class="portrait">FLIP TO PLAY</span>
+                </div>
 
-    <!-- Background Image as CSS class -->
+                <!-- Round Countdown Overlay -->
+                <div class="countdown-overlay d-none">
+                    <div class="countdown ready mb-5"></div>
+                    <!-- <span class="count">3</span> -->
+                </div>
 
-    <!-- Play Again Button -->
-    <div class="play-again-container position-absolute d-flex justify-content-center align-items-end">
-        <div id="pulse" class="play-again mb-5"></div>
-    </div>
+                <!-- Background Image as CSS class -->
 
-    <!-- Game Result Banner -->
-    <div class="result-container position-absolute d-flex justify-content-center">
-        <div class="result-banner position-absolute mt-4"></div>
-    </div>
+                <!-- Play Again Button -->
+                <div class="play-again-container position-absolute d-flex justify-content-center align-items-end">
+                    <div id="pulse" class="play-again mb-5"></div>
+                </div>
+
+                <!-- Game Result Banner -->
+                <div class="result-container position-absolute d-flex justify-content-center">
+                    <div class="result-banner position-absolute mt-4"></div>
+                </div>
 
 
-    <!-- Timer Counter -->
-    <div class="timer-counter">
-        <!-- Transitioning between two decimals to one vice versa using d-none (display none class utility in Bootstrap),
+                <!-- Timer Counter -->
+                <div class="timer-counter">
+                    <!-- Transitioning between two decimals to one vice versa using d-none (display none class utility in Bootstrap),
      being controlled in JavaScript frontend. -->
-        <!-- <div class="time-no-decimal number-6 d-none"></div>
+                    <!-- <div class="time-no-decimal number-6 d-none"></div>
         <div class="time-left-decimal number-1"></div>
         <div class="time-right-decimal number-0"></div> -->
-        <span class="time-nums">30</span>
-    </div>
+                    <span class="time-nums">30</span>
+                </div>
 
                 <!-- Main Components -->
                 <div class="container-fluid">
@@ -209,8 +208,7 @@
 
                     <!-- Characters' Pigeon Components -->
                     <div class="pigeons-container position-absolute">
-                        <img src="/assets/img/pigeons/pigeonV1/PigeonV1-9.png" alt=""
-                            class="pigeon-left ">
+                        <img src="/assets/img/pigeons/pigeonV1/PigeonV1-9.png" alt="" class="pigeon-left ">
                         <img src="/assets/img/pigeons/pigeonV3/PigeonV3-1.png" alt="" class="pigeon-right">
                     </div>
 
@@ -244,6 +242,7 @@
         </div>
         <div class="row">
             <div class=" bottom-background">
+                <img src="/assets/img/LYPSTYC-logo.png" alt="LIPSTYC Logo" class="companyLogo">
             </div>
         </div>
 </body>
